refactor: moved to auth route for now removed signin route
